requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.1.0
aiohttp>=3.9.1
asyncio>=3.4.3
python-dotenv>=1.0.0
tqdm>=4.66.1
lxml>=4.9.3
